Extract aggregation logic
